lightning-polar-scripts/pyproject.toml
[tool.poetry]
name = "lightning-polar-scripts"
version = "0.1.0"
description = "Scripts for interacting with Polar Lightning Network (Core Lightning and LND)"
authors = ["bitwalt <walterm21@proton.me>"]
readme = "README.md"
packages = [{include = "lightning_polar_scripts"}]

[tool.poetry.dependencies]
python = "^3.11"
# HTTP client for REST APIs
requests = "^2.31.0"
# CLI and utilities
click = "^8.1.7"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^23.12.0"
isort = "^5.13.0"
flake8 = "^6.1.0"
mypy = "^1.8.0"

[tool.poetry.scripts]
polar-cli = "lightning_polar_scripts.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
